Add files via upload
These files are for the python version of the program, however doing this project in python turned out to be a silly idea when it could be done much more easily in java! Here is what I accomplished in python, I will be uploading the final Java program (no BASH files will need to be used if all goes well) here as well and that will be fully functional.